extract LbmDomain into own module
